Removed additional file.
